🌱 Split NotificationSettingsSection and ActiveAlerts into focused sub-components
Extract self-contained sub-components to reduce hook count and line count
in two high-complexity components:

NotificationSettingsSection (456 → ~95 lines):
- BrowserNotificationSettings: permission request + verification flow
- SlackNotificationSettings: webhook config + test notification
- EmailNotificationSettings: SMTP config + test notification

ActiveAlerts (498 → ~260 lines):
- AlertListItem: individual alert row rendering + quick actions
- NotificationVerifyIndicator: inline browser notification verification

Partially addresses #2500
