+++ web/src/components/cards/ActiveAlerts.tsx
@@ -1,48 +1,25 @@
 import { useState, useMemo, useCallback } from 'react'
 import {
   AlertTriangle,
-  Bell,
   CheckCircle,
-  Clock,
-  ChevronRight,
-  Bot,
-  Server,
   Eye,
   EyeOff,
-  ExternalLink,
+  Server,
 } from 'lucide-react'
 import { useAlerts } from '../../hooks/useAlerts'
 import { StatusBadge } from '../ui/StatusBadge'
 import { useGlobalFilters, type SeverityLevel } from '../../hooks/useGlobalFilters'
 import { useDrillDown } from '../../hooks/useDrillDown'
 import { useMissions } from '../../hooks/useMissions'
-import { getSeverityIcon } from '../../types/alerts'
 import type { Alert, AlertSeverity } from '../../types/alerts'
-import { Button } from '../ui/Button'
 import { CardControls } from '../ui/CardControls'
 import { Pagination } from '../ui/Pagination'
-import { useCardData, CardClusterFilter, CardSearchInput, CardAIActions } from '../../lib/cards'
+import { useCardData, CardClusterFilter, CardSearchInput } from '../../lib/cards'
 import { useCardLoadingState } from './CardDataContext'
 import { useTranslation } from 'react-i18next'
-import type { TFunction } from 'i18next'
 import { useDemoMode } from '../../hooks/useDemoMode'
-import { isBrowserNotifVerified, setBrowserNotifVerified } from '../../lib/notificationStatus'
-
-// Severity color map — defined at module level to avoid re-creation on each render
-const SEVERITY_COLORS: Record<AlertSeverity, string> = {
-  critical: 'bg-red-500/20 text-red-400 border-red-500/30',
-  warning: 'bg-orange-500/20 text-orange-400 border-orange-500/30',
-  info: 'bg-blue-500/20 text-blue-400 border-blue-500/30',
-}
-
-// Severity indicator badge — extracted to module level to prevent re-creation on every render
-function SeverityBadge({ severity }: { severity: AlertSeverity }) {
-  return (
-    <span className={`px-1.5 py-0.5 text-xs rounded border ${SEVERITY_COLORS[severity]}`}>
-      {severity}
-    </span>
-  )
-}
+import { NotificationVerifyIndicator } from './NotificationVerifyIndicator'
+import { AlertListItem } from './AlertListItem'
 
 // Stats summary row shown at the top of the alerts card
 function AlertStatsRow({ critical, warning, acknowledged }: { critical: number; warning: number; acknowledged: number }) {
@@ -74,23 +51,8 @@ function AlertStatsRow({ critical, warning, acknowledged }: { critical: number; 
   )
 }
 
-// Format relative time
-function formatRelativeTime(dateString: string, t: TFunction<'cards'>): string {
-  const date = new Date(dateString)
-  const now = new Date()
-  const diffMs = now.getTime() - date.getTime()
-  const diffMins = Math.floor(diffMs / 60000)
-  const diffHours = Math.floor(diffMins / 60)
-  const diffDays = Math.floor(diffHours / 24)
-
-  if (diffMins < 1) return t('activeAlerts.justNow')
-  if (diffMins < 60) return t('activeAlerts.minutesAgo', { count: diffMins })
-  if (diffHours < 24) return t('activeAlerts.hoursAgo', { count: diffHours })
-  return t('activeAlerts.daysAgo', { count: diffDays })
-}
-
-/** Notification verification flow state */
-type NotifVerifyState = 'idle' | 'asked' | 'verified' | 'failed'
+/** Default pagination size for the alerts list */
+const DEFAULT_PAGE_SIZE = 5
 
 type SortField = 'severity' | 'time'
 
@@ -110,42 +72,6 @@ export function ActiveAlerts() {
   const { missions, setActiveMission, openSidebar } = useMissions()
 
   const [showAcknowledged, setShowAcknowledged] = useState(false)
-
-  // Browser notification verification state
-  const [notifVerifyState, setNotifVerifyState] = useState<NotifVerifyState>('idle')
-  const [notifVerified, setNotifVerified] = useState(() => isBrowserNotifVerified())
-
-  /** Whether the notification indicator should be visible */
-  const showNotifIndicator =
-    typeof Notification !== 'undefined' &&
-    Notification.permission === 'granted' &&
-    !notifVerified &&
-    notifVerifyState !== 'verified'
-
-  /** Send a test notification and ask user to confirm receipt */
-  const handleSendTestNotif = () => {
-    try {
-      new Notification('KubeStellar Console', {
-        body: t('activeAlerts.testNotificationBody'),
-        icon: '/favicon.ico',
-      })
-    } catch {
-      // Notification constructor may throw in some environments
-    }
-    setNotifVerifyState('asked')
-  }
-
-  /** User confirmed they saw the test notification */
-  const handleNotifYes = () => {
-    setBrowserNotifVerified(true)
-    setNotifVerified(true)
-    setNotifVerifyState('verified')
-  }
-
-  /** User did NOT see the test notification */
-  const handleNotifNo = () => {
-    setNotifVerifyState('failed')
-  }
 
   // Combine active and acknowledged alerts when toggle is on
   const allAlertsToShow = useMemo(() => {
@@ -237,7 +163,7 @@ export function ActiveAlerts() {
         time: (a, b) => new Date(b.firedAt).getTime() - new Date(a.firedAt).getTime(),
       },
     },
-    defaultLimit: 5,
+    defaultLimit: DEFAULT_PAGE_SIZE,
   })
 
   const handleAlertClick = (alert: Alert) => {
@@ -292,40 +218,8 @@ export function ActiveAlerts() {
               {localClusterFilter.length}/{availableClustersForFilter.length}
             </span>
           )}
-          {/* Subtle browser notification verification indicator */}
-          {showNotifIndicator && notifVerifyState === 'idle' && (
-            <button
-              onClick={handleSendTestNotif}
-              title={t('activeAlerts.notifNotVerified')}
-              className="relative flex items-center p-1 rounded hover:bg-secondary/60 transition-colors"
-            >
-              <Bell className="w-3.5 h-3.5 text-muted-foreground" />
-              <span className="absolute -top-0.5 -right-0.5 w-2 h-2 rounded-full bg-amber-400" />
-            </button>
-          )}
-          {notifVerifyState === 'asked' && (
-            <span className="flex items-center gap-1.5 text-xs text-muted-foreground">
-              <span>{t('activeAlerts.didYouSeeIt')}</span>
-              <button
-                onClick={handleNotifYes}
-                className="px-1.5 py-0.5 rounded bg-green-500/20 text-green-400 hover:bg-green-500/30 transition-colors text-xs"
-              >
-                {t('activeAlerts.yes')}
-              </button>
-              <button
-                onClick={handleNotifNo}
-                className="px-1.5 py-0.5 rounded bg-secondary text-muted-foreground hover:text-foreground transition-colors text-xs"
-              >
-                {t('activeAlerts.no')}
-              </button>
-            </span>
-          )}
-          {notifVerifyState === 'failed' && (
-            <span className="flex items-center gap-1 text-xs text-amber-400">
-              <Bell className="w-3 h-3" />
-              <span>{t('activeAlerts.checkSystemSettings')}</span>
-            </span>
-          )}
+          {/* Browser notification verification indicator */}
+          <NotificationVerifyIndicator />
         </div>
 
         <div className="flex items-center gap-2">
@@ -393,87 +287,15 @@ export function ActiveAlerts() {
           </div>
         ) : (
           displayedAlerts.map((alert: Alert) => (
-            <div
+            <AlertListItem
               key={alert.id}
-              onClick={() => handleAlertClick(alert)}
-              className="p-2 rounded-lg bg-secondary/30 border border-border/50 hover:bg-secondary/50 cursor-pointer transition-colors group"
-            >
-              <div className="flex items-start gap-2">
-                <span className="text-lg">{getSeverityIcon(alert.severity)}</span>
-                <div className="flex-1 min-w-0">
-                  <div className="flex items-center gap-2 mb-1">
-                    <span className="text-sm font-medium text-foreground truncate">
-                      {alert.ruleName}
-                    </span>
-                    <SeverityBadge severity={alert.severity} />
-                  </div>
-                  <p className="text-xs text-muted-foreground line-clamp-2">
-                    {alert.message}
-                  </p>
-                  <div className="flex items-center gap-3 mt-1.5">
-                    {alert.cluster && (
-                      <span className="text-xs text-muted-foreground flex items-center gap-1">
-                        <Server className="w-3 h-3" />
-                        {alert.cluster}
-                      </span>
-                    )}
-                    <span className="text-xs text-muted-foreground flex items-center gap-1">
-                      <Clock className="w-3 h-3" />
-                      {formatRelativeTime(alert.firedAt, t)}
-                    </span>
-                    {getMissionForAlert(alert) && (
-                      <span className="text-xs text-purple-400 flex items-center gap-1">
-                        <Bot className="w-3 h-3" />
-                        AI
-                      </span>
-                    )}
-                    {alert.acknowledgedAt && (
-                      <span className="text-xs text-green-400">{t('activeAlerts.acknowledged')}</span>
-                    )}
-                  </div>
-                </div>
-                <ChevronRight className="w-4 h-4 text-muted-foreground opacity-0 group-hover:opacity-100 transition-opacity" />
-              </div>
-
-              {/* Quick Actions */}
-              <div className="flex items-center gap-2 mt-2 pt-2 border-t border-border/30">
-                {!alert.acknowledgedAt && (
-                  <Button
-                    variant="secondary"
-                    size="sm"
-                    onClick={e => handleAcknowledge(e, alert.id)}
-                    className="rounded"
-                  >
-                    {t('activeAlerts.acknowledge')}
-                  </Button>
-                )}
-                {(() => {
-                  const mission = getMissionForAlert(alert)
-                  if (mission) {
-                    return (
-                      <Button
-                        variant="accent"
-                        size="sm"
-                        onClick={e => handleOpenMission(e, alert)}
-                        icon={<ExternalLink className="w-3 h-3" />}
-                        className="rounded"
-                      >
-                        {t('activeAlerts.viewDiagnosis')}
-                      </Button>
-                    )
-                  } else {
-                    return (
-                      <CardAIActions
-                        resource={{ kind: 'Alert', name: alert.ruleName, cluster: alert.cluster, status: alert.severity }}
-                        issues={[{ name: alert.ruleName, message: alert.message }]}
-                        showRepair={false}
-                        onDiagnose={e => handleAIDiagnose(e, alert.id)}
-                      />
-                    )
-                  }
-                })()}
-              </div>
-            </div>
+              alert={alert}
+              mission={getMissionForAlert(alert)}
+              onAlertClick={handleAlertClick}
+              onAcknowledge={handleAcknowledge}
+              onAIDiagnose={handleAIDiagnose}
+              onOpenMission={handleOpenMission}
+            />
           ))
         )}
       </div>
@@ -485,7 +307,7 @@ export function ActiveAlerts() {
             currentPage={currentPage}
             totalPages={totalPages}
             totalItems={totalItems}
-            itemsPerPage={typeof itemsPerPage === 'number' ? itemsPerPage : 5}
+            itemsPerPage={typeof itemsPerPage === 'number' ? itemsPerPage : DEFAULT_PAGE_SIZE}
             onPageChange={goToPage}
             showItemsPerPage={false}
           />
